Add files via upload
This is a basic and simple veichle controller in Unity.  You can just import the .unitypackage in to your scene.

This controller uses WheelsColliders and components from this built-inUnity tool. The code is not originaly mine, I've followed the tutorial of the great @b3agz for a enviroment art project, and I decided to put up in here in case I develop another project that includes veichles movements.

If you use the code, don't forget to change the names of the meshes and the colliders to the same names it is referenced in the code, and most important, please tag and give credit to @b3agz.